Moved middleware functions to own package
